Move generateTask and answer_handler to src
